Fix cirrupt NPC bug and add blank NPC names
Using NPC memory 0x08 to create a unique ID. This appears safe, but if errors are found, another word location can be used (requires changing line 1113). This will not corrupt existing projects.

Also added a condition to check for a blank string in the dialogue name. This will allow blank names without a stray ":".